+++ src/Modules/CleanupModule.php
@@ -62,7 +62,5 @@ final class CleanupModule implements ModuleInterface
         remove_action('wp_head', 'rsd_link');
         remove_action('wp_head', 'wlwmanifest_link');
         remove_action('wp_head', 'wp_shortlink_wp_head', 10);
-        remove_action('wp_head', 'rest_output_link_wp_head', 10);
-        remove_action('wp_head', 'wp_oembed_add_discovery_links', 10);
     }
 }